+++ test/compare.js
@@ -1,7 +1,7 @@
 'use strict';
 var tic = process.hrtime();
 
-var percentile = require('../index');
+var percentile = require('../percentile');
 var scipyPercentile = require('./scipy_quantile');
 var fs = require('fs');
 var path = require('path');
